카메라 코드 분리
